All the moderation stuff
Here is all the moderation stuff for the node classifier and diffusion, as well as some visualisations. Everything is run from moderation_driver. So just cd to moderation and then run "py moderation_driver.py". You will probably need to setup the file paths for the csvs that is needed to run it. You can set them at the bottom of the moderation_driver file. At the top you can also alter thresholds if you wanted to run the node classification moderation with different settings.