Moderation Implementation & Visualisations (#2)
* DCRNN works.

* Better result.

* Not sure what it does, but works.

* Archival.

* Run/debug config.

* At least duncitonal.

* All of this works.

* Balanced the score bins.

* Ignore csv and pycache.

* Loss function alternatives.

* More experiments.

* Bare PyG GCN.

* V2 refactored stack, first cut.

* Trash.

* Add attention to a static GCN.

* WIP.

* Base GCN + Logits loss.

* More tuning.

* Refactor.

* Removed evaluation mode.

* Everythng runs

* Cleaned up documentation.

* No more ds store/

* Include tabulate.

* Threshold cross validation.

* Adjust thresholding.

* mode = "classification"

* Cross validation.

* CV.

* Functional static gcn.

* Misc cleanup.

* Decent results.

* Cross validation.

* Launch JSON config.

* Early stopping

* Best resulsts.

* All the moderation stuff

Here is all the moderation stuff for the node classifier and diffusion, as well as some visualisations. Everything is run from moderation_driver. So just cd to moderation and then run "py moderation_driver.py". You will probably need to setup the file paths for the csvs that is needed to run it. You can set them at the bottom of the moderation_driver file. At the top you can also alter thresholds if you wanted to run the node classification moderation with different settings.

---------

Co-authored-by: Sujay <sujayanjankar@gmail.com>